Prompt before deleting a page, and fix observers.
We can't update text when Core Data sends a change notification because it messes up the editor. Instead, we should only update the editor if we receive a notification that we've received a change from iCloud. At that point, a cloud update always wins over whatever we happen to be typing. LocalDB update notifications trigger syntax highlighting for links, but nothing else. We can assume that whats in the editor is what's in the database. That's how it got in the database in the first place.